Move communication management module to shared dir
Refactor into functional style for consistency.
